pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = "Template for django project"
authors = ["Dmytriy Kovaliov <monkdok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
psycopg2 = "^2.9.5"
pre-commit = "^3.2.0"
flake8-bugbear = "^23.3.23"
flake8-builtins = "^2.1.0"
flake8-comprehensions = "^3.11.1"
flake8-print = "^5.0.0"
flake8-quotes = "^3.3.2"
psycopg2-binary = "^2.9.5"
black = "^23.1.0"
factory-boy = "^3.2.1"

[tool.poetry.group.prod]
optional = true

[tool.poetry.group.prod.dependencies]
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
